function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pandas
azure.kusto.data
pyodbc
sqlalchemy
manufacturingmetrics-0.1.3-py3-none-any.whl